requirements.txt
# Core framework
crewai==0.130.0
crewai-tools==0.47.1

# API
fastapi==0.110.3
uvicorn==0.29.0
python-multipart==0.0.9

# LLM - let pip resolve openai version automatically (crewai needs >=1.68.2)
openai>=1.68.2
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# PDF processing
pypdf==4.2.0

# Utilities
python-dotenv==1.0.1
pydantic>=2.7.1

# Optional: Bonus queue worker support
redis==5.0.4
celery==5.3.6